requirements.txt
altgraph==0.17
asgiref==3.3.1
image==1.5.33
importlib-metadata==3.4.0
macholib==1.14
Pillow==8.1.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
